:construction: POC rabbitmq setup
