+++ .gitignore
@@ -20,6 +20,9 @@
 # secrets
 *.env
 
+# Data and shit
+data/
+
 # Please output directory and local configuration
 plz-out
 .plzconfig.local
