+++ Dockerfile-send
@@ -13,12 +13,12 @@ RUN go mod download
 COPY . .
 
 # Build the Go app
-RUN CGO_ENABLED=0 GOOS=linux go build -o crawler cmd/crawler/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -o send cmd/rabbitmq-tutorial/send/main.go
 
 
 ######## Start a new stage from scratch #######
 FROM scratch
 
-COPY --from=builder /build/crawler /crawler
+COPY --from=builder /build/send /send
 
-CMD ["./crawler"]
+CMD ["./send"]
