+++ go.mod
@@ -7,4 +7,5 @@ require (
 	github.com/gocolly/colly/v2 v2.1.0
 	github.com/hashicorp/golang-lru v0.5.4
 	github.com/lib/pq v1.8.0
+	github.com/streadway/amqp v1.0.0
 )
